requirements.txt
hikari
hikari-lightbulb
python-dotenv
asyncio
lightbulb-ext-tungsten